reduced cache and refactored to work same
